feat: add terraform for ec2/tailscale
Add the terraform (OpenTofu) necessary for managing my EC2 instances and
Tailscale (to the extent supported by the Tailscale provider).

The main terraform module is `ec2-homelab` which provisions an ec2
instance and configures a username and boots/auths tailscale (using a
pre-generated key). The remainder of the provisioning will occur via
Ansible.

We create the `dev` and `global` Terraform environments even though they
are not yet utilized.

Add the shell scripts for Terraform bootstrapping.

A future todo is adding `shellcheck` to our pre-commit (and running it
on existing scripts).