feat: add open-webui app
Add the first part of the `open-webui` application. This commit contains
an overall description of the application and the configuration files
(Caddy and compose.yml) for running the application on any Docker
Compose host).

It took quite a bit of tweaking to get Caddyfile and the tailscale
sidecar working. We need to ensure we run tailscale in user mode,
because the host instance will also be running tailscale, and we don't
want them to conflict.

All of the containers share a single network, and refer to each other
over localhost. The tailscale container is the single point for
controlling what's exposed to the "outside" world.

In a follow-up commit, we will add the Ansible configuration for
deploying this docker-compose/Caddyfile/.env combo.