feat: add initial ansible playbooks for host configuration
We plan to use ansible for the initial configuring of our hosts.
The main alternative would be going purely declarative with Nix, but
that's not something I'm interested in right now.

These playbooks are intended to be run _after_ the host is already
accessible via Tailscale. On an Ec2 instance, the `user_data.sh` should
be able to get us to that point. For an on-prem machine, it'll require
some manual action.

These playbooks should be self explanatory. They are all safe run
repeatedly, with the possible exception of the `btrfs` playbook.

We use `uv` for managing our python environment.

We delegate to localhost for fetching secrets, etc from AWS Secrets Manager. This
approach let's us avoid installing boto3 on the target machine, and
let's us manage secret access in a more centralized way (i.e. we don't
need to give every remote machine access to AWS Secrets Manager. Doing
so would've been fine for EC2 instances, but a real hassle for on-prem
machines).

By default, we target our playbooks at all hosts, unless there's a
specific need to limit to a more targeted group (i.e. the `nvidia-gpu`
playbook).

We've not yet defined any ansible roles. We _may_ later.

In general, we will targetoy all of our software via containers
(either docker-compose or k8s). We don't anticipate ever writing a
Ansible playbook to install software on bare-metal.