feat: deploy open-webui onto EC2
In this commit, we introduce a new playbook called
`docker-compose-tailscale-service`, and then utilize the playbook to
deploy `open-webui`.

`docker-compose-tailscale-service` is simple. It loads the requested
secrets from aws-secrets-manager. Then, it copies over the
files/templates in the pertinent `{{ service_name }}` subdirectory, and
then restarts `docker compose` if anything changed.

We discussed the configuration for `open-webui` in a commit for
`software/apps/open-webui`.

For now, we do not symlink to the `software/apps/open-webui` directory.
We may want to eventually.

This commit is an initial draft. It's not perfect. We still have updates
to make to the docker compose, and we want to better understand how we
survive restarts, etc.