wip: feat: deploy open-webui onto EC2
In this commit, we introduce a new playbook called
`docker-compose-tailscale-service`, and then utilize the playbook to
deploy `open-webui`.

`docker-compose-tailscale-service` is simple. It loads the requested
secrets from aws-secrets-manager. Then, it copies over the
files/templates in the pertinent `{{ service_name }}` subdirectory, and
then restarts `docker compose` if anything changed.

We discussed the configuration for `open-webui` in a commit for
`software/apps/open-webui`.

For now, we do not symlink to the `software/apps/open-webui` directory.
We may want to eventually.

TODO:

For us to merge this commit, we want to do the following:

- Set-up the `@data` btrfs subvolume (and tbd also create additional
  @data/@postgres, @data/@open-webui), etc... for storing Docker data.
  Mount as directories.
- Verify that `open-webui` survives a full `site.yml` run.
- Verify that `open-webui` survives machine restart.
- Get Bedrock Access Gateway and a Postgres side-car container working.