feat: switch open-webui to on-prem and start exploring lxd on-prem
I'll deploy open-webui on-prem. The singular advantage to the EC2
instance was being able to use IAM instance profile for controlling
Bedrock access, but that's not worth the cost of running the instance. I
plan to move on-prem. Delete the actual prod instance of open-webui.
We've applied this `terraform`, so we're no longer paying for the
instance. We've not yet deployed open-webui on-prem, but we're moving
that direction.

Also make updates to the `docker-compose-tailscale-service` playbook
(particularly around managing btrfs subvolumes and directory mounts).
We want docker "base" to use Ext4 and then we will mount in BtrFS
directories as necessary.

Include some early experiments w/ lxd VMs. Nothing permanent yet.