feat: add traefik-proxy
Utilize https://github.com/hectorm/docker-proxy to make a docker-proxy
available. With the approach outlines in `poc`, we can control all
outbound traffic from our containers.

We've _not yet_ applied this to the main `compose.yaml`.

We opted to copy-paste the upstream code (which is MIT licensed) instead
of using a git submodule. The code is sufficiently simple that the cost
of "maintaining" is minimal, and I want to keep dependencies to a
minimal.

We've included the LICENSE.md, which applies to the software in this
directory only.