feat(ai): add LiteLLM for model access
Add LiteLLM (backed by Postgres DB) as new AI proxy. With this approach,
we can access any variety of models via the OpenAI spec.

- Configure with support for Anthropic and Bedrock models (Llama, Titan,
  etc)
- Set up Tailscale as ephemeral to survive container restarts
- Create, and use, AWS service user module with cost monitoring, and use
  access/secret key from service user for AWS access. We don't _love_
  using the hardcoded credentials, but it should be okay in the
  short-term. We've restricted the permissions to Bedrock, set
  cost-alerting, and the .env file is on an encrypted disk. Longer term,
  we can look into IP allowlisting the credentials and only storing them
  in memory (i.e. never on disk).

We've verified this approach survives a `docker compose up/down/up`
cycle. Importantly, we need to NOT store any tailscaled state outside of
the container and we need to set `?ephemeral=true` when running
`tailscale up`. Otherwise, `tailscale` will get confused about whether
to create a new node or not, and connections weren't working.

We remove a legacy open-webui set-up which we will replace later.