feat: deploy open-webui
Deploy open-webui! Overall, pretty straightforward. Our configuration
focuses on privacy.

We have _stopped_ including `port` command in the tailscale container.
It's not necessary and it causes conflict.

We also update our `hosts.yml` file so that we can target `open-webui`
and `litellm` separately.

Finally, we increased the size of our vm-docker-compose VM.